editted and added mult/u

+++ src/cpu.cpp
@@ -190,7 +190,7 @@ void CPU::execute_r_type(R_Instruction inst) {
             advance_pc(4);
             break;
         case OpFunction::MULT: {
-            Product product = get_register(inst.src1) * get_register(inst.src2);
+            Product product = static_cast<int64_t>(get_register(inst.src1)) * static_cast<int64_t>(get_register(inst.src2));
             LO = product & 0xFFFFFFFF;
             HI = (product >> 32) & 0xFFFFFFFF;
             advance_pc(4);
